+++ qgis/providers/grass/qgsgrassprovider.h
@@ -172,7 +172,7 @@ public:
 
   /** Return the extent for this data layer
    */
-  virtual QgsRect *extent(int dataSourceLayerNum = 0);
+  virtual QgsRect *extent(size_t dataSourceLayerNum = 0);
 
   /**
    * Get the field information for the layer
