+++ qgis/providers/postgres/qgspostgresprovider.cpp
@@ -778,7 +778,7 @@ void QgsPostgresProvider::setExtent( QgsRect* newExtent )
 }
 
 // TODO - make this function return the real extent_
-QgsRect *QgsPostgresProvider::extent(int dataSourceLayerNum)
+QgsRect *QgsPostgresProvider::extent(size_t dataSourceLayerNum)
 {
   return &layerExtent;      //extent_->MinX, extent_->MinY, extent_->MaxX, extent_->MaxY);
 }
